src/speer-notes/requirements.txt
Eve==2.2.0
Flask-Cors==5.0.0
requests==2.32.3
hypermea-core==0.9.38

hypermea==0.9.38
Flask~=3.1.0
Cerberus~=1.3.7
pymongo~=4.11.2

# start: added by add-validation
isodate==0.7.2
# end: added by add-validation

# start: added by add-auth
eve-negotiable-auth==0.9.6
PyJWT==2.10.1
cryptography==44.0.2
# end: added by add-auth

# start: added by add-redis
redis==5.2.1
# end: added by add-redis

# start: added by add-test
pytest~=8.3.5
pytest-bdd==8.1.0
assertpy==1.1
# end: added by add-test
